+++ src/lib/waku_relay/index.ts
@@ -17,6 +17,7 @@ import { InMessage } from 'libp2p-interfaces/src/pubsub';
 import { SignaturePolicy } from 'libp2p-interfaces/src/pubsub/signature-policy';
 import PeerId from 'peer-id';
 
+import { hexToBuf } from '../utils';
 import { CreateOptions, DefaultPubSubTopic } from '../waku';
 import { WakuMessage } from '../waku_message';
 
@@ -64,9 +65,6 @@ export class WakuRelay extends Gossipsub {
   heartbeat: RelayHeartbeat;
   pubSubTopic: string;
 
-  /**
-   * Decryption private keys to use to attempt decryption of incoming messages.
-   */
   public decryptionKeys: Set<Uint8Array>;
 
   /**
@@ -124,21 +122,24 @@ export class WakuRelay extends Gossipsub {
   }
 
   /**
-   * Register a decryption private key or symmetric key to attempt decryption
-   * of messages received on the given content topic. This can either be a
-   * private key for asymmetric encryption or a symmetric key. Waku relay will
-   * attempt to decrypt messages using both methods.
+   * Register a decryption key to attempt decryption of received messages.
+   * This can either be a private key for asymmetric encryption or a symmetric
+   * key. `WakuRelay` will attempt to decrypt messages using both methods.
+   *
+   * Strings must be in hex format.
    */
-  addDecryptionKey(privateKey: Uint8Array): void {
-    this.decryptionKeys.add(privateKey);
+  addDecryptionKey(key: Uint8Array | string): void {
+    this.decryptionKeys.add(hexToBuf(key));
   }
 
   /**
-   * Delete a decryption key to attempt decryption of messages received on the
-   * given content topic.
+   * Delete a decryption key that was used to attempt decryption of received
+   * messages.
+   *
+   * Strings must be in hex format.
    */
-  deleteDecryptionKey(privateKey: Uint8Array): void {
-    this.decryptionKeys.delete(privateKey);
+  deleteDecryptionKey(key: Uint8Array | string): void {
+    this.decryptionKeys.delete(hexToBuf(key));
   }
 
   /**